HeapSort, PriorityQueue, BinaryTreeDepthLists
Added heapsort based on my array backed heap

Added a method that returns a list of lists of elements(T). One list for each 'row' in the binary tree

Implemented a priority queue using my queue implementation and a sorted dictionary. Im not sure whether it should derive from queue or not. It can do all the things a queue can and functions like a queue so in that way it makes sense but it doesn't make use of the inner list of queue and it overrides all the functions so it doesnt benefit much from being derived from it.